Further display adjustments
Mostly involving wresting with getting the list of commits on to one
line and being well behaved. I've had to add the 'width: 80vw' in order
to keep the width under control. There must be a better way to handle
this but I've not been able to figure out it.

We also individually color the links so that we can make the author
links a different colour. Not sure of this is better but it at least it
introduces a little visible variety.

Also remove the grey background from the repo name on the 'new commits'
lines in order to try to get them to fade into the background a little
more.